Move TraceSource engine-side as agentic-traces schema
TraceSource keys the trace-replay harnesses the engine dispatches on, so
the keys belong to the engine (same split as WorkflowVenvType in the
venv-provisioner PR): it moves to llm_module/agentic_traces/schema.py
and is re-exported from reference_config's agentic_traces_config, which
keeps the per-model content (run specs, mode settings, git-ref pins).

runs.py now imports TraceSource from the schema module and narrows its
AgenticTracesConfig/AgenticTracesRunSpec imports to TYPE_CHECKING (they
are adapter config types the engine only reads structurally);
agentic_traces_tests.py takes TraceSource from the engine module and its
config lookups already go through the TargetPack. Adapter callers
(workflow_venvs, validate_setup, tests) keep working via the re-export.
